+++ packages/avatar/src/element.ts
@@ -1,14 +1,6 @@
-import {
-  LitElement,
-  CSSResult,
-  TemplateResult,
-  PropertyValues,
-  html,
-  property,
-  queryAsync,
-  eventOptions,
-} from 'lit-element';
-import { ifDefined } from 'lit-html/directives/if-defined';
+import { LitElement, CSSResult, TemplateResult, PropertyValues, html } from 'lit';
+import { property, queryAsync, eventOptions } from 'lit/decorators';
+import { ifDefined } from 'lit/directives/if-defined';
 import Picture from '@equinor/fusion-wc-picture';
 import Ripple, { RippleHandlers } from '@equinor/fusion-wc-ripple';
 import style from './element.css';
@@ -56,7 +48,7 @@ export class AvatarElement extends LitElement implements AvatarElementProps {
     return this.ripple;
   });
 
-  protected updated(changedProperties: PropertyValues) {
+  protected updated(changedProperties: PropertyValues): void {
     super.updated(changedProperties);
     if (changedProperties.has('disabled')) {
       if (this.disabled) {
@@ -68,11 +60,13 @@ export class AvatarElement extends LitElement implements AvatarElementProps {
   }
 
   protected renderPicture(): TemplateResult {
-    return html`<fwc-picture class="picture" src=${ifDefined(this.src)} cover></fwc-picture> `;
+    return html`<div class="fwc-avatar__picture-container">
+      <fwc-picture class="fwc-avatar__picture" src=${ifDefined(this.src)} cover></fwc-picture>
+    </div>`;
   }
 
   protected renderValue(): TemplateResult {
-    return html`${this.value}`;
+    return html`<span class="fwc-avatar__value">${this.value}</span>`;
   }
 
   protected renderSlot(): TemplateResult {
@@ -82,7 +76,7 @@ export class AvatarElement extends LitElement implements AvatarElementProps {
   protected render(): TemplateResult {
     const content = this.src ? this.renderPicture() : this.value ? this.renderValue() : this.renderSlot();
     return html`<span
-      class="circle"
+      class="fwc-avatar__container"
       @focus="${this.handleRippleFocus}"
       @blur="${this.handleRippleBlur}"
       @mousedown="${this.handleRippleActivate}"
@@ -97,12 +91,12 @@ export class AvatarElement extends LitElement implements AvatarElementProps {
 
   protected renderRipple(): TemplateResult | string {
     return this.clickable
-      ? html`<fwc-ripple class="ripple" disabled="${ifDefined(this.disabled)}" unbounded></fwc-ripple>`
+      ? html`<fwc-ripple class="fwc-avatar__ripple" disabled="${ifDefined(this.disabled)}" unbounded></fwc-ripple>`
       : '';
   }
 
   @eventOptions({ passive: true })
-  protected handleRippleActivate(evt?: Event) {
+  protected handleRippleActivate(evt?: Event): void {
     const onUp = () => {
       window.removeEventListener('mouseup', onUp);
 
@@ -113,23 +107,23 @@ export class AvatarElement extends LitElement implements AvatarElementProps {
     this.rippleHandlers.startPress(evt);
   }
 
-  protected handleRippleDeactivate() {
+  protected handleRippleDeactivate(): void {
     this.rippleHandlers.endPress();
   }
 
-  protected handleRippleMouseEnter() {
+  protected handleRippleMouseEnter(): void {
     this.rippleHandlers.startHover();
   }
 
-  protected handleRippleMouseLeave() {
+  protected handleRippleMouseLeave(): void {
     this.rippleHandlers.endHover();
   }
 
-  protected handleRippleFocus() {
+  protected handleRippleFocus(): void {
     this.rippleHandlers.startFocus();
   }
 
-  protected handleRippleBlur() {
+  protected handleRippleBlur(): void {
     this.rippleHandlers.endFocus();
   }
 }