+++ packages/avatar/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.18](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-avatar@1.0.17...@equinor/fusion-wc-avatar@1.0.18) (2021-09-23)
+
+
+### Bug Fixes
+
+* fixed linting ([ca6a86e](https://github.com/equinor/fusion-web-components/commit/ca6a86ebda14f6c85cb58f125778e94847b70b1d))
+* upgraded lit-element and lit-html to lit 2.0.0 with other packages ([93cd2f9](https://github.com/equinor/fusion-web-components/commit/93cd2f997d6045fd5ab69fe05ccee5acfa861ad7))
+
+
+
+
+
+## [1.0.17](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-avatar@1.0.16...@equinor/fusion-wc-avatar@1.0.17) (2021-09-22)
+
+
+### Bug Fixes
+
+* minor avatar changes to styling and naming conventions according to person avatar requirements ([f3d99af](https://github.com/equinor/fusion-web-components/commit/f3d99af81e478ac65a987cc33a0928add645fbcc))
+
+
+
+
+
+## [1.0.16](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-avatar@1.0.15...@equinor/fusion-wc-avatar@1.0.16) (2021-09-21)
+
+**Note:** Version bump only for package @equinor/fusion-wc-avatar
+
+
+
+
+
 ## [1.0.15](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-avatar@1.0.14...@equinor/fusion-wc-avatar@1.0.15) (2021-09-20)