+++ storybook/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.4](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-storybook@1.9.3...@equinor/fusion-wc-storybook@1.9.4) (2021-09-23)
+
+
+### Bug Fixes
+
+* upgraded lit-element and lit-html to lit 2.0.0 with other packages ([93cd2f9](https://github.com/equinor/fusion-web-components/commit/93cd2f997d6045fd5ab69fe05ccee5acfa861ad7))
+
+
+
+
+
+## [1.9.3](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-storybook@1.9.2...@equinor/fusion-wc-storybook@1.9.3) (2021-09-22)
+
+**Note:** Version bump only for package @equinor/fusion-wc-storybook
+
+
+
+
+
+## [1.9.2](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-storybook@1.9.1...@equinor/fusion-wc-storybook@1.9.2) (2021-09-21)
+
+**Note:** Version bump only for package @equinor/fusion-wc-storybook
+
+
+
+
+
 ## [1.9.1](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-storybook@1.9.0...@equinor/fusion-wc-storybook@1.9.1) (2021-09-20)